refactor(theming): add type definitions for the theme object
The theme object contains mostly CSS rules, but only some are supported/configurable. This will add
a type definition for the theme object to allow ease of use when customizing the theme, providing a
clear definition of what can be edited and what kind of CSS rule values are supported.

fix #10